# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SecondLine Integration API
  description: >
    The SecondLine (Atheris Control Solution) side of the ThirdLine / NexusRisk
    integration contract (BRD §8). A control is defined once and consumed
    everywhere: control definitions, test results with effectiveness ratings,
    and open exceptions are published so internal audit can place reliance on
    second-line testing. All endpoints are scoped to the tenant that owns the
    presented API key. Writes require an Idempotency-Key header; repeated
    delivery of the same key is acknowledged without reprocessing.
  version: 1.0.0
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: secondline.client.example
security:
  - ApiKeyAuth: []
paths:
...
  /frameworks/coverage:
    get:
      summary: Framework coverage headline per framework
      description: >
        Share of each framework's testable requirements covered by an approved
        control mapping, and the share where the mapped control also carries a
        current effective rating. Unapproved mappings and retired controls are
        excluded. Global frameworks plus any the tenant owns.
      parameters:
        - name: jurisdiction
          in: query
          description: ISO-3166 alpha-2 code, or INTL
          schema: { type: string }
      responses:
        '200':
          description: Coverage per framework
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/FrameworkCoverage' }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
  schemas:
...
    FrameworkCoverage:
      type: object
      properties:
        framework_id: { type: integer }
        code: { type: string, example: COSO-IC-2013 }
        name: { type: string }
        testable_requirements: { type: integer }
        mapped_requirements: { type: integer, description: At least one approved mapping to an active control }
        effective_requirements: { type: integer, description: Mapped control also carries a current effective rating }
        unmapped_requirements: { type: integer }
        mapped_percentage: { type: integer }
        effective_percentage: { type: integer }
        verification_status: { type: string, enum: [verified, unverified, draft] }
...